Let Web Installer skip DB setup with bundled MySQL.
If you choose to use the bundled MySQL database, we already know the
password for the root user, the database type, and so on.

We can populate this information in the installer.json file so
it will be picked up by the OTRS web installer automatically.
This makes sure that end users don't need to specify any
username or password for MySQL if they choose to use the bundled
database.

The only 'difficult' thing is that we need to pre-populate
the password for the 'otrs' database account that will be created
by the web isntaller. For this we use the NSIS Pwgen plug-in.